+++ frontend/src/constants/convertSupportedFornats.ts
@@ -1,0 +1,21 @@
+// Central list of formats supported by Convert operations
+export const CONVERT_SUPPORTED_FORMATS = [
+  // Microsoft Office
+  'doc', 'docx', 'dot', 'dotx', 'csv', 'xls', 'xlsx', 'xlt', 'xltx', 'slk', 'dif', 'ppt', 'pptx',
+  // OpenDocument
+  'odt', 'ott', 'ods', 'ots', 'odp', 'otp', 'odg', 'otg',
+  // Text formats
+  'txt', 'text', 'xml', 'rtf', 'html', 'lwp', 'md',
+  // Images
+  'bmp', 'gif', 'jpeg', 'jpg', 'png', 'tif', 'tiff', 'pbm', 'pgm', 'ppm', 'ras', 'xbm', 'xpm', 'svg', 'svm', 'wmf', 'webp',
+  // StarOffice
+  'sda', 'sdc', 'sdd', 'sdw', 'stc', 'std', 'sti', 'stw', 'sxd', 'sxg', 'sxi', 'sxw',
+  // Email formats
+  'eml',
+  // Archive formats
+  'zip',
+  // Other
+  'dbf', 'fods', 'vsd', 'vor', 'vor3', 'vor4', 'uop', 'pct', 'ps', 'pdf',
+];
+
+
