+++ frontend/src/contexts/ToolWorkflowContext.tsx
@@ -13,123 +13,18 @@ import { useNavigationUrlSync } from '../hooks/useUrlSync';
 import { getDefaultWorkbench } from '../types/workbench';
 import { filterToolRegistryByQuery } from '../utils/toolSearch';
 import { useToolHistory } from '../hooks/tools/useToolHistory';
-import { FullscreenToolStyleSettings, defaultFullscreenToolSettings } from '../components/tools/FullscreenToolSettings';
+import { FullscreenToolStyleSettings } from '../components/tools/FullscreenToolSettings';
+import {
+  ToolWorkflowState,
+  TOOL_PANEL_MODE_STORAGE_KEY,
+  FULLSCREEN_TOOL_SETTINGS_STORAGE_KEY,
+  createInitialState,
+  toolWorkflowReducer,
+  ToolPanelMode,
+} from './toolWorkflow/state';
 
 // State interface
-type ToolPanelMode = 'sidebar' | 'fullscreen';
-
-interface ToolWorkflowState {
-  // UI State
-  sidebarsVisible: boolean;
-  leftPanelView: 'toolPicker' | 'toolContent' | 'hidden';
-  readerMode: boolean;
-  toolPanelMode: ToolPanelMode;
-  fullscreenToolSettings: FullscreenToolStyleSettings;
-
-  // File/Preview State
-  previewFile: File | null;
-  pageEditorFunctions: PageEditorFunctions | null;
-
-  // Search State
-  searchQuery: string;
-}
-
-// Actions
-type ToolWorkflowAction =
-  | { type: 'SET_SIDEBARS_VISIBLE'; payload: boolean }
-  | { type: 'SET_LEFT_PANEL_VIEW'; payload: 'toolPicker' | 'toolContent' | 'hidden' }
-  | { type: 'SET_READER_MODE'; payload: boolean }
-  | { type: 'SET_TOOL_PANEL_MODE'; payload: ToolPanelMode }
-  | { type: 'SET_FULLSCREEN_TOOL_SETTINGS'; payload: FullscreenToolStyleSettings }
-  | { type: 'SET_PREVIEW_FILE'; payload: File | null }
-  | { type: 'SET_PAGE_EDITOR_FUNCTIONS'; payload: PageEditorFunctions | null }
-  | { type: 'SET_SEARCH_QUERY'; payload: string }
-  | { type: 'RESET_UI_STATE' };
-
-// Initial state
-export const TOOL_PANEL_MODE_STORAGE_KEY = 'toolPanelModePreference';
-export const FULLSCREEN_TOOL_SETTINGS_STORAGE_KEY = 'fullscreenToolStyleSettings';
-export const LEGACY_TOOL_SETTINGS_STORAGE_KEY = 'legacyToolStyleSettings';
-
-const getStoredToolPanelMode = (): ToolPanelMode => {
-  if (typeof window === 'undefined') {
-    return 'sidebar';
-  }
-
-  const stored = window.localStorage.getItem(TOOL_PANEL_MODE_STORAGE_KEY);
-  if (stored === 'legacy' || stored === 'fullscreen') {
-    return 'fullscreen';
-  }
-
-  return 'sidebar';
-};
-
-const getStoredFullscreenToolSettings = (): FullscreenToolStyleSettings => {
-  if (typeof window === 'undefined') {
-    return defaultFullscreenToolSettings;
-  }
-
-  try {
-    const storedNew = window.localStorage.getItem(FULLSCREEN_TOOL_SETTINGS_STORAGE_KEY);
-    if (storedNew) {
-      return { ...defaultFullscreenToolSettings, ...JSON.parse(storedNew) };
-    }
-    const storedLegacy = window.localStorage.getItem(LEGACY_TOOL_SETTINGS_STORAGE_KEY);
-    if (storedLegacy) {
-      return { ...defaultFullscreenToolSettings, ...JSON.parse(storedLegacy) };
-    }
-  } catch (e) {
-    console.error('Failed to parse fullscreen tool settings:', e);
-  }
-
-  return defaultFullscreenToolSettings;
-};
-
-const baseState: Omit<ToolWorkflowState, 'toolPanelMode' | 'fullscreenToolSettings'> = {
-  sidebarsVisible: true,
-  leftPanelView: 'toolPicker',
-  readerMode: false,
-  previewFile: null,
-  pageEditorFunctions: null,
-  searchQuery: '',
-};
-
-const createInitialState = (): ToolWorkflowState => ({
-  ...baseState,
-  toolPanelMode: getStoredToolPanelMode(),
-  fullscreenToolSettings: getStoredFullscreenToolSettings(),
-});
-
-// Reducer
-function toolWorkflowReducer(state: ToolWorkflowState, action: ToolWorkflowAction): ToolWorkflowState {
-  switch (action.type) {
-    case 'SET_SIDEBARS_VISIBLE':
-      return { ...state, sidebarsVisible: action.payload };
-    case 'SET_LEFT_PANEL_VIEW':
-      return { ...state, leftPanelView: action.payload };
-    case 'SET_READER_MODE':
-      return { ...state, readerMode: action.payload };
-    case 'SET_TOOL_PANEL_MODE':
-      return { ...state, toolPanelMode: action.payload };
-    case 'SET_FULLSCREEN_TOOL_SETTINGS':
-      return { ...state, fullscreenToolSettings: action.payload };
-    case 'SET_PREVIEW_FILE':
-      return { ...state, previewFile: action.payload };
-    case 'SET_PAGE_EDITOR_FUNCTIONS':
-      return { ...state, pageEditorFunctions: action.payload };
-    case 'SET_SEARCH_QUERY':
-      return { ...state, searchQuery: action.payload };
-    case 'RESET_UI_STATE':
-      return {
-        ...baseState,
-        toolPanelMode: state.toolPanelMode,
-        fullscreenToolSettings: state.fullscreenToolSettings,
-        searchQuery: state.searchQuery,
-      };
-    default:
-      return state;
-  }
-}
+// Types and reducer/state moved to './toolWorkflow/state'
 
 // Context value interface
 interface ToolWorkflowContextValue extends ToolWorkflowState {
@@ -266,7 +161,6 @@ export function ToolWorkflowProvider({ children }: ToolWorkflowProviderProps) {
 
     const serialized = JSON.stringify(state.fullscreenToolSettings);
     window.localStorage.setItem(FULLSCREEN_TOOL_SETTINGS_STORAGE_KEY, serialized);
-    window.localStorage.setItem(LEGACY_TOOL_SETTINGS_STORAGE_KEY, serialized);
   }, [state.fullscreenToolSettings]);
 
   // Tool reset methods
